implement permission on app/Http/Controllers/UserController.php and app/Http/Controllers/WalletController.php

+++ app/Http/Controllers/UserController.php
@@ -7,6 +7,12 @@ use Illuminate\Http\Request;
 
 class UserController extends Controller
 {
+    private $request;
+
+    public function __construct(Request $request)
+    {
+        $this->request = $request;
+    }
     /**
      * Display a listing of the resource.
      *
@@ -14,16 +20,20 @@ class UserController extends Controller
      */
     public function index(Request $request)
     {
-        $users = User::paginate(20);
-        return $this->ok('Get users success', $users);
+        if ($request->can('SHOW_USER')) {
+            $users = User::paginate(20);
+            return $this->ok('Get users success', $users);
+        } else {
+            return $this->unauthorized();
+        }
     }
 
     public function profile(Request $request)
     {
         $user = $request->user();
-        if($user->can('SHOW_USER_PROFILE')){
+        if ($user->can('SHOW_USER_PROFILE')) {
             return $this->ok('Get profile success', $user);
-        }else{
+        } else {
             return $this->unauthorized();
         }
     }
@@ -37,15 +47,18 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $input = $request->all();
-        $user = User::find('id');
-        if (!$user) {
-            return $this->badRequest('User id ' . $id . ' not found');
+        if ($request->can('UPDATE_USER')) {
+            $input = $request->all();
+            $user = User::find('id');
+            if (!$user) {
+                return $this->badRequest('User id ' . $id . ' not found');
+            }
+            $user->fill($input);
+            $user->save();
+            return $this->save('Update user success', $user);
+        } else {
+            return $this->unauthorized();
         }
-
-        $user->fill($input);
-        $user->save();
-        return $this->save('Update user success', $user);
     }
 
     /**
@@ -56,12 +69,16 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $user = User::find('id');
-        if (!$user) {
-            return $this->badRequest('User id ' . $id . ' not found');
-        }
+        if ($this->request->can('DESTROY_USER')) {
+            $user = User::find('id');
+            if (!$user) {
+                return $this->badRequest('User id ' . $id . ' not found');
+            }
 
-        $user->delete();
-        return $this->save('Update user success', $user);
+            $user->delete();
+            return $this->save('Update user success', $user);
+        } else {
+            return $this->unauthorized();
+        }
     }
 }
